blog-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pyjwt==2.8.0
pydantic==2.5.3
pydantic-settings==2.1.0
